+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## Unreleased
+### Fixed
+- Filter even on else now uses the config to set FilterBuilder.
 
 ## 0.4.0 (3. September, 2023)
 ### Changed